Crate ptr_union

Pointer union types the size of a pointer by storing the tag in the alignment bits.

Structs

Builder2

A builder for Union2.

Builder4

A builder for Union4.

Builder8

A builder for Union8.

Builder16

A builder for Union16.

Union2

A pointer union of two pointer types.

Union4

A pointer union of four pointer types.

Union8

A pointer union of eight pointer types.

Union16

A pointer union of up to sixteen pointer types.

Enums

Enum2

An unpacked version of Union2.

Enum4

An unpacked version of Union4.

Enum8

An unpacked version of Union4.

Enum16

An unpacked version of Union8.