ptr_union
Enum Enum16

pub enum Enum16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P> {

    A(A),
    B(B),
    C(C),
    D(D),
    E(E),
    F(F),
    G(G),
    H(H),
    I(I),
    J(J),
    K(K),
    L(L),
    M(M),
    N(N),
    O(O),
    P(P),

}

An unpacked version of Union8.

Variants

A(A)

B(B)

C(C)

D(D)

E(E)

F(F)

G(G)

H(H)

I(I)

J(J)

K(K)

L(L)

M(M)

N(N)

O(O)

P(P)

Implementations

impl<A: ErasablePtr, B: ErasablePtr, C: ErasablePtr, D: ErasablePtr, E: ErasablePtr, F: ErasablePtr, G: ErasablePtr, H: ErasablePtr, I: ErasablePtr, J: ErasablePtr, K: ErasablePtr, L: ErasablePtr, M: ErasablePtr, N: ErasablePtr, O: ErasablePtr, P: ErasablePtr> Enum16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

pub fn pack( self, builder: Builder16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Pack this loose enum into a pointer union.

pub fn try_pack( self, ) -> Result<Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>, Self>

Pack this loose enum into a pointer union.

pub unsafe fn pack_unchecked( self, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Pack this loose enum into a pointer union.

Safety

The pointer packed must be properly aligned. Note that only the actually packed pointer is restricted, not any other pointer type involved in this definition.