Struct ptr_union::Builder2

pub struct Builder2<A, B> { /* fields omitted */ }

A builder for [Union2].

An instance of this builder means that Union2 parameterized with the same type arguments are safe to construct.

Implementations

impl<A, B> Builder2<A, B>

pub const unsafe fn new_unchecked() -> Self

Assert that creating pointer unions of these types is safe.

Safety

The pointer types must be erasable, and their alignment must meet the requirements of the target union type.

impl<A: ErasablePtr, B: ErasablePtr> Builder2<A, B>

pub fn a(self, this: A) -> Union2<A, B>

Construct a union at this variant.

pub fn b(self, this: B) -> Union2<A, B>

Construct a union at this variant.