ptr_union
Struct Builder16

pub struct Builder16<A, B, C, D, E, F = NeverPtr, G = NeverPtr, H = NeverPtr, I = NeverPtr, J = NeverPtr, K = NeverPtr, L = NeverPtr, M = NeverPtr, N = NeverPtr, O = NeverPtr, P = NeverPtr> { /* private fields */ }

A builder for Union16.

An instance of this builder means that Union16 parameterized with the same type arguments are safe to construct.

Implementations

impl<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P> Builder16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

pub const unsafe fn new_unchecked() -> Self

Assert that creating pointer unions of these types is safe.

Safety

The pointer types must be erasable, and their alignment must meet the requirements of the target union type.

impl<A: ErasablePtr, B: ErasablePtr, C: ErasablePtr, D: ErasablePtr, E: ErasablePtr, F: ErasablePtr, G: ErasablePtr, H: ErasablePtr, I: ErasablePtr, J: ErasablePtr, K: ErasablePtr, L: ErasablePtr, M: ErasablePtr, N: ErasablePtr, O: ErasablePtr, P: ErasablePtr> Builder16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

pub fn a( self, this: A, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn b( self, this: B, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn c( self, this: C, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn d( self, this: D, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn e( self, this: E, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn f( self, this: F, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn g( self, this: G, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn h( self, this: H, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn i( self, this: I, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn j( self, this: J, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn k( self, this: K, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn l( self, this: L, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn m( self, this: M, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn n( self, this: N, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn o( self, this: O, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.

pub fn p( self, this: P, ) -> Union16<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P>

Construct a union at this variant.